move mailbox_dashboard_controller to controller folder
move mailbox_dashboard_controller to controller folder and add search_controller to mailbox_dashbroad_binding

(cherry picked from commit 313a3dad7e7201d869edfd05e829ed1bd69c7385)
